LocalChannel支持windows平台

+++ hterm/channel_local.py
@@ -1,11 +1,16 @@
 import os
-import pty
-import fcntl
-import struct
-import termios
+import sys
 import threading
 
 from channel_pty import PtyChannel
+
+if sys.platform == 'win32':
+    import pywinpty
+else:
+    import pty
+    import fcntl
+    import struct
+    import termios
 
 
 class LocalChannel(PtyChannel):
@@ -16,10 +21,13 @@ class LocalChannel(PtyChannel):
         self._running = True
     
     def connect(self) -> None:
-        self.pid, self.fd = pty.fork()
-        if self.pid == 0:
-            # 子进程执行目标程序
-            os.execvp(self.progname, [self.progname])
+        if sys.platform == 'win32':
+            self.proc = PtyProcess.spawn(self.progname)
+        else:
+            self.pid, self.fd = pty.fork()
+            if self.pid == 0:
+                # 子进程执行目标程序
+                os.execvp(self.progname, [self.progname])
         self.is_connected = True
         self.connected.emit(f'\r\nLocalChannel: {self.progname} connected.\r\n')
         
@@ -32,21 +40,34 @@ class LocalChannel(PtyChannel):
         self.is_connected = False
         self.disconnected.emit("User requested disconnect.")
         self._running = False
-        os.close(self.fd)
-        self._thread.join()
-        os.kill(self.pid, termios.SIGHUP)
-        os.waitpid(self.pid, os.WNOHANG)
+        if sys.platform == 'win32':
+            self.proc.close(force=True)
+            self._thread.join()
+        else:
+            os.close(self.fd)
+            self._thread.join()
+            os.kill(self.pid, termios.SIGHUP)
+            os.waitpid(self.pid, os.WNOHANG)
 
     def send_data(self, data: str) -> None:
-        data_bytes = data.encode()
-        os.write(self.fd, data_bytes)
+        if sys.platform == 'win32':
+            self.proc.write(data)
+        else:
+            data_bytes = data.encode()
+            os.write(self.fd, data_bytes)
         
-    def send_window_size(self, rows: int, cols: int) -> None:
-        winsize = struct.pack('HHHH', rows, cols, 0, 0)
-        fcntl.ioctl(self.fd, termios.TIOCSWINSZ, winsize)
+    def send_window_size(self, rows: int, cols: int):
+        if sys.platform == 'win32':
+            self.proc.setwinsize(rows, cols)
+        else:
+            winsize = struct.pack('HHHH', rows, cols, 0, 0)
+            fcntl.ioctl(self.fd, termios.TIOCSWINSZ, winsize)
 
     def receive_loop(self):
         while self._running:
-            data_bytes = os.read(self.fd, 1024)
-            data_str = data_bytes.decode("utf-8", "replace")
+            if sys.platform == 'win32':
+                data_str = self.proc.read()
+            else:
+                data_bytes = os.read(self.fd, 1024)
+                data_str = data_bytes.decode("utf-8", "replace")
             self.received.emit(data_str)
